Fix an issue that promise was not called / handled
Problems happen "only" in the common case that
login was called and there is no Digits session.
When the new activity was open and the callback
was called before the react-native app is shown
the application never receives the promise solve
handler .then().
With an delay, or, in this case with an Lifecycle
listener which waits until "onResume" was called
this works fine.